+++ executables/application/src/PlaceShipControllerView.cpp
@@ -11,7 +11,7 @@ namespace cpp_warships::application {
     void PlaceShipControllerView::displayShipsLeft() const {
         std::cout << "All available ships (length : count)" << std::endl;
 
-        int emptyTypes = 0;
+        size_t emptyTypes = 0;
         for (const auto& pair : controller->getAvailableLengthShips()) {
             if (pair.second != 0) {
                 std::cout << "Ship of length " << pair.first << " is needed to place " << pair.second
